requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
boto3==1.24.96
botocore==1.27.96
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-allauth==0.51.0
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
dynaconf==3.1.11
h11==0.14.0
hiredis==2.0.0
hvac==1.0.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==21.3.0
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.6.0
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.2
Pillow==9.2.0
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.5.0
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==21.1.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
txaio==22.2.1
types-cryptography==3.3.23.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
validators==0.20.0
Werkzeug==2.2.2
whitenoise==6.2.0
win32-setctime==1.1.0
zipp==3.9.0
zope.interface==5.4.0
django-webshell
